+++ UnityProject/Saneject/Assets/Tests/Runtime/MethodInjection/IDependency.cs
@@ -1,0 +1,7 @@
+﻿namespace Tests.Runtime.MethodInjection
+{
+    public interface IDependency
+    {
+        
+    }
+}