+++ UnityProject/Saneject/Assets/Tests/Runtime/MethodInjection/MyClassWithDeepNested.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using UnityEngine;
+
+namespace Tests.Runtime.MethodInjection
+{
+    public class MyClassWithDeepNested : MonoBehaviour
+    {
+        [SerializeField]
+        private MyNestedForDeep nestedClass;
+    }
+
+    [Serializable]
+    public class MyNestedForDeep
+    {
+        [SerializeField]
+        private MyDeepNestedClass deepNested;
+    }
+}